feat: Add GHES support (#23)
This PR adds support for generating user and team links when VSCode is
configured with a Github Enterprise Server authentication provider, as
suggested in #22 .

One consequence is links to users and teams will use the GHES url when
switching to a repo that was cloned from github.com. This is an issue
any plugin will face in this scenario and not unique to this one. In
situations like this, it may be advisable for users to utilize VSCode's
Workspace feature, keeping GHES-based repos in one Workspace and
Github.com-based repos in another while syncing their extensions and
settings appropriately.